# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UrbanEye",
    "description": "Evidence-first community intelligence starter",
    "version": "0.1.0"
  },
  "paths": {
...
    "/areas/{geo_id}": {
      "get": {
        "summary": "Get Area",
        "operationId": "get_area_areas__geo_id__get",
        "parameters": [
          {
            "name": "geo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Area"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/areas/{geo_id}/history": {
      "get": {
        "summary": "Get Area History",
        "operationId": "get_area_history_areas__geo_id__history_get",
        "parameters": [
          {
            "name": "geo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geo Id"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "median_household_income",
              "title": "Metric"
            }
          },
          {
            "name": "span",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Span"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/areas/{geo_id}/changes": {
      "get": {
        "summary": "Get Area Changes",
        "operationId": "get_area_changes_areas__geo_id__changes_get",
        "parameters": [
          {
            "name": "geo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geo Id"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "median_household_income",
              "title": "Metric"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2021,
              "title": "From Year"
            }
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2024,
              "title": "To Year"
            }
          },
          {
            "name": "span",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Span"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/areas/{geo_id}/businesses": {
      "get": {
        "summary": "Get Area Businesses",
        "operationId": "get_area_businesses_areas__geo_id__businesses_get",
        "parameters": [
          {
            "name": "geo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/compare": {
      "get": {
        "summary": "Get Compare",
        "operationId": "get_compare_compare_get",
        "parameters": [
          {
            "name": "geo_ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geo Ids"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2024,
              "title": "Year"
            }
          },
          {
            "name": "metrics",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "median_household_income",
              "title": "Metrics"
            }
          },
          {
            "name": "span",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Span"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ask": {
      "post": {
        "summary": "Ask",
        "operationId": "ask_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/segment": {
      "post": {
        "summary": "Segment",
        "operationId": "segment_segment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluate": {
      "post": {
        "summary": "Evaluate",
        "description": "Business site evaluation: LLM-generated strengths/concerns grounded in public evidence.",
        "operationId": "evaluate_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Answer": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "mode": {
            "type": "string",
            "enum": [
              "facts",
              "retrieval",
              "insufficient_evidence",
              "llm"
            ],
            "title": "Mode"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "summary",
          "limitations",
          "evidence_ids",
          "evidence"
        ],
        "title": "Answer"
      },
      "Area": {
        "properties": {
          "geo_id": {
            "type": "string",
            "minLength": 1,
            "title": "Geo Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "geography_type": {
            "type": "string",
            "enum": [
              "census_designated_place",
              "census_tract",
              "block_group"
            ],
            "title": "Geography Type"
          },
          "boundary_vintage": {
            "type": "string",
            "title": "Boundary Vintage"
          },
          "metrics": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Metrics"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "geo_id",
          "name",
          "geography_type",
          "boundary_vintage",
          "metrics",
          "evidence"
        ],
        "title": "Area"
      },
...
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Question"
          },
          "geo_id": {
            "type": "string",
            "minLength": 1,
            "title": "Geo Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "geo_id"
        ],
        "title": "AskRequest"
      },
...
      "BusinessResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "geo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Category"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/BusinessPoi"
            },
            "type": "array",
            "title": "Features"
          },
          "attribution": {
            "type": "string",
            "title": "Attribution"
          },
          "license": {
            "type": "string",
            "title": "License"
          },
          "copyright_url": {
            "type": "string",
            "title": "Copyright Url"
          },
          "source_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Date"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total",
          "by_category",
          "features",
          "attribution",
          "license",
          "copyright_url",
          "limitations"
        ],
        "title": "BusinessResponse"
      },
      "ChangeResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "geo_id": {
            "type": "string",
            "title": "Geo Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "geography_type": {
            "type": "string",
            "title": "Geography Type"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "span": {
            "type": "integer",
            "title": "Span"
          },
          "from_year": {
            "type": "integer",
            "title": "From Year"
          },
          "to_year": {
            "type": "integer",
            "title": "To Year"
          },
          "from_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Estimate"
          },
          "to_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Estimate"
          },
          "absolute_change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Change"
          },
          "percent_change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percent Change"
          },
          "inflation_adjusted": {
            "type": "boolean",
            "title": "Inflation Adjusted",
            "default": false
          },
          "significant_90": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Significant 90"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "geo_id",
          "name",
          "geography_type",
          "metric",
          "span",
          "from_year",
          "to_year",
          "from_estimate",
          "to_estimate",
          "absolute_change",
          "percent_change",
          "limitations"
        ],
        "title": "ChangeResponse"
      },
...
      "CompareResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metrics"
          },
          "areas": {
            "items": {
              "$ref": "#/components/schemas/CompareArea"
            },
            "type": "array",
            "title": "Areas"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "year",
          "metrics",
          "areas",
          "limitations"
        ],
        "title": "CompareResponse"
      },
...
      "EvaluateRequest": {
        "properties": {
          "geo_id": {
            "type": "string",
            "minLength": 1,
            "title": "Geo Id"
          },
          "business_type": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Business Type",
            "description": "Type of business being considered, e.g. 'coffee shop'."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "geo_id",
          "business_type"
        ],
        "title": "EvaluateRequest",
        "description": "Business site evaluation request. business_type is a short user-supplied label."
      },
      "EvaluateResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "geo_id": {
            "type": "string",
            "title": "Geo Id"
          },
          "area_name": {
            "type": "string",
            "title": "Area Name"
          },
          "business_type": {
            "type": "string",
            "title": "Business Type"
          },
          "mode": {
            "type": "string",
            "enum": [
              "evaluated",
              "insufficient_evidence",
              "llm_unavailable"
            ],
            "title": "Mode"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/EvaluateFinding"
            },
            "type": "array",
            "title": "Strengths"
          },
          "concerns": {
            "items": {
              "$ref": "#/components/schemas/EvaluateFinding"
            },
            "type": "array",
            "title": "Concerns"
          },
          "customer_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Context"
          },
          "competition_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competition Context"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "geo_id",
          "area_name",
          "business_type",
          "mode",
          "summary",
          "evidence_ids",
          "evidence",
          "limitations"
        ],
        "title": "EvaluateResponse",
        "description": "LLM-powered business site brief. Every claim cites supplied evidence."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "geo_id": {
            "type": "string",
            "title": "Geo Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "geography_type": {
            "type": "string",
            "title": "Geography Type"
          },
          "zcta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zcta"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "span": {
            "type": "integer",
            "title": "Span"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistorySource"
              },
              {
                "type": "null"
              }
            ]
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/HistoryPoint"
            },
            "type": "array",
            "title": "Series"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "geo_id",
          "name",
          "geography_type",
          "metric",
          "span",
          "series",
          "limitations"
        ],
        "title": "HistoryResponse"
      },
...
      "SegmentRequest": {
        "properties": {
          "geo_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 2,
            "title": "Geo Ids"
          },
          "x_metric": {
            "type": "string",
            "minLength": 1,
            "title": "X Metric"
          },
          "y_metric": {
            "type": "string",
            "minLength": 1,
            "title": "Y Metric"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "geo_ids",
          "x_metric",
          "y_metric"
        ],
        "title": "SegmentRequest"
      },
      "SegmentResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "x_metric": {
            "type": "string",
            "title": "X Metric"
          },
          "y_metric": {
            "type": "string",
            "title": "Y Metric"
          },
          "correlation_coefficient": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Coefficient"
          },
          "sample_size": {
            "type": "integer",
            "title": "Sample Size"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "data_points": {
            "items": {
              "$ref": "#/components/schemas/DataPoint"
            },
            "type": "array",
            "title": "Data Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x_metric",
          "y_metric",
          "correlation_coefficient",
          "sample_size",
          "explanation",
          "data_points"
        ],
        "title": "SegmentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}